[C] Implement C raw ping pong sample.
